Task: Move EventModel Enum from eventModel to CommonModel
- Move EventCode enum from eventModel to CommonEventModel
- Move EventCode String() method from eventModel to CommonEventModel
